Permutations Of A string
In a non-programmatic way, we generally swap the last two characters and fix the rest of the characters in their respective positions. Then, we move to the previous character (i.e ){n-3, n->length of input string } and so on by decreasing it by 1. So,at each level, we get the 
 nth permutation by swapping with the indices starting from the respective start index up to the length of the string For example, if the word is "abcd", I start at index  0 (remains at the call stack,number of possible combo : 4),then moves to 1st index (remains at the call stack, number of possible combo : 3) followed by index 2 and index 3 with respective combinations 2 and 1. so if I'm able to print out all those combinations,it would be 4x3x2x1=24  Here we go.


Time Complexity: O(n*n!)